requirements.txt
colorama >= 0.3.7
lxml >= 3.6.0
ndg-httpsclient >= 0.4.0
pex >= 1.1.6
py2exe_py2 >= 0.6.9
pygal >= 2.2.3
wheel >= 0.29.0
ushlex >= 0.99
